cleanup decision id enricher
