+++ test/unit/enricher/decisions.spec.ts
@@ -1,5 +1,5 @@
 import { expect, use } from 'chai'
-import * as decisions from '../../../src/manager/decisions'
+import { enrichDecisionIds } from '../../../src/enrichers/decisions'
 import { DefaultStorageHandler } from 'live-connect-handlers'
 // @ts-expect-error
 import uuid from 'tiny-uuid4'
@@ -18,6 +18,8 @@ describe('DecisionsManager for stored decisions', () => {
   let externalStorage: DefaultStorageHandler
   let storage: WrappedStorageHandler
 
+  const domain = 'example.com'
+
   beforeEach(() => {
     eventBus = LocalEventBus()
     externalStorage = new DefaultStorageHandler(eventBus)
@@ -29,26 +31,26 @@ describe('DecisionsManager for stored decisions', () => {
   })
 
   it('should return an empty string if nothing is in the cookie jar', () => {
-    const resolutionResult = decisions.resolve({}, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain })
     expect(resolutionResult.decisionIds).to.eql([])
   })
 
   it('should return an empty string if the cookie jar has invalid uuids', () => {
     storage.setCookie('lidids.', '2134')
-    const resolutionResult = decisions.resolve({}, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain })
     expect(resolutionResult.decisionIds).to.eql([])
   })
 
   it('should return the stored decision', () => {
     const decisionId = uuid()
     storage.setCookie('lidids.123', decisionId)
-    const resolutionResult = decisions.resolve({}, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain })
     expect(resolutionResult.decisionIds).to.eql([decisionId])
   })
 
   it('should not return empty values', () => {
     storage.setCookie('lidids.', '')
-    const resolutionResult = decisions.resolve({}, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain })
     expect(resolutionResult.decisionIds.length).to.eql(0)
   })
 
@@ -56,7 +58,7 @@ describe('DecisionsManager for stored decisions', () => {
     const decisionId = uuid()
     storage.setCookie('lidids.123', decisionId, undefined, undefined, 'something.example.com')
     storage.setCookie('lidids.123', decisionId, undefined, undefined, 'www.something.example.com')
-    const resolutionResult = decisions.resolve({}, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain })
     expect(resolutionResult.decisionIds).to.eql([decisionId])
   })
 })
@@ -66,6 +68,8 @@ describe('DecisionsManager for new decisions', () => {
   let eventBus: EventBus
   let externalStorage: DefaultStorageHandler
   let storage: WrappedStorageHandler
+
+  const domain = 'example.com'
 
   beforeEach(() => {
     sandbox = sinon.createSandbox()
@@ -80,21 +84,21 @@ describe('DecisionsManager for new decisions', () => {
 
   it('should return the new decision id, and store it', () => {
     const decisionId = uuid()
-    const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionId}` }, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionId}` })
     expect(resolutionResult.decisionIds).to.eql([decisionId])
     expect(storage.getCookie(`lidids.${decisionId}`)).to.eq(decisionId)
   })
 
   it('should return the new decision id, and store it', () => {
     const decisionId = uuid()
-    const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionId}` }, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionId}` })
     expect(resolutionResult.decisionIds).to.eql([decisionId])
     expect(storage.getCookie(`lidids.${decisionId}`)).to.eq(decisionId)
   })
 
   it('should not return the new decision id if its not a uuid', () => {
     const decisionId = `${uuid()}sometghing`
-    const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionId}` }, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionId}` })
     expect(resolutionResult.decisionIds).to.eql([])
   })
 
@@ -102,7 +106,7 @@ describe('DecisionsManager for new decisions', () => {
     const decisionIdOne = uuid()
     const decisionIdTwo = uuid()
     storage.setCookie('lidids.123', decisionIdOne)
-    const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=    ${decisionIdTwo}` }, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=    ${decisionIdTwo}` })
     expect(resolutionResult.decisionIds).to.eql([decisionIdTwo, decisionIdOne])
     expect(storage.getCookie(`lidids.${decisionIdTwo}`)).to.eq(decisionIdTwo)
   })
@@ -112,7 +116,7 @@ describe('DecisionsManager for new decisions', () => {
     const decisionIdTwo = uuid()
     const decisionIdThree = uuid()
     storage.setCookie('lidids.123', decisionIdOne)
-    const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionIdTwo}&li_did=${decisionIdThree}` }, storage, eventBus)
+    const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionIdTwo}&li_did=${decisionIdThree}` })
     expect(resolutionResult.decisionIds).to.eql([decisionIdTwo, decisionIdThree, decisionIdOne])
     expect(storage.getCookie(`lidids.${decisionIdTwo}`)).to.eq(decisionIdTwo)
     expect(storage.getCookie(`lidids.${decisionIdThree}`)).to.eq(decisionIdThree)
@@ -133,7 +137,7 @@ describe('DecisionsManager for new decisions', () => {
       stub => stub.restore(),
       () => {
         expect(errorEmitted).to.eq(false)
-        const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionIdTwo}&li_did=${decisionIdThree}` }, storage, eventBus)
+        const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionIdTwo}&li_did=${decisionIdThree}` })
         expect(resolutionResult.decisionIds).to.eql([decisionIdTwo, decisionIdThree])
         expect(storage.getCookie(`lidids.${decisionIdTwo}`)).to.eq(decisionIdTwo)
         expect(storage.getCookie(`lidids.${decisionIdThree}`)).to.eq(decisionIdThree)
@@ -156,7 +160,7 @@ describe('DecisionsManager for new decisions', () => {
       stub => stub.restore(),
       () => {
         expect(errorEmitted).to.eq(false)
-        const resolutionResult = decisions.resolve({ pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionIdTwo}&li_did=${decisionIdThree}` }, storage, eventBus)
+        const resolutionResult = enrichDecisionIds(storage, eventBus)({ domain, pageUrl: `http://subdomain.tests.example.com/cake?li_did=${decisionIdTwo}&li_did=${decisionIdThree}` })
         expect(resolutionResult.decisionIds).to.eql([decisionIdTwo, decisionIdThree, decisionIdOne])
         expect(errorEmitted).to.eq(true)
       })
